Engine now exits gracefully
Major change to the engine to gracefully exit. No more exit(0) functions
are used, using exceptions, and errors are caught and sent back to Java
without crashing the android application. This allows quicker return back
to the android app and static memory to stay in memory.

Also added a new function callback for ONScripterView to get notified when
finished is called either when user exits the app or it crashes.