Add support for saving on sdcard using treeUri
Uses the api for DocumentFile.fromTreeUri to read save files. Your app
will need to save sdcard paths as persisted uris so that if a game's
path is from the sdcard it will be able to save files. If there is no
permissions and game exists on sdcard, it will error.

Also saves gloval.sav when onPause occurs so save files that will be
shown on load screens for some games if user decides to swipe away.